modularize useActionForm
